requirements.txt
asgiref==3.5.2
attrs==22.1.0
Babel==2.11.0
beautifulsoup4==4.11.1
Django==4.1.3
django-bootstrap4==22.2
django-phonenumber-field==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.17.0
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
uritemplate==4.1.1
